django_audio_manager/requirements.txt
Django==2.2.8
psycopg2==2.8.4
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
